# openapi/mediapilot-core-api.yaml
openapi: 3.0.3

info:
  title: MediaFlow Core API
  version: 1.0.0
  description: Core API for managing personal media library items.

servers:
  - url: http://localhost:8080
    description: Local Spring Boot backend

tags:
  - name: Media
    description: Media management endpoints

## MEDIA
paths:
  /api/media:
    get:
      tags:
        - MediaQuery
      operationId: findAllMedia
      summary: Find all media
      responses:
        '200':
          description: List of media
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './components/media-schemas.yaml#/components/schemas/MediaResponse'

    post:
      tags:
        - MediaCommand
      operationId: createMedia
      summary: Create a media item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/media-schemas.yaml#/components/schemas/CreateMediaRequest'
      responses:
        '201':
          description: Media created
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'

  /api/media/{mediaId}:
    get:
      tags:
        - MediaQuery
      operationId: findMediaById
      summary: Find media by id
      responses:
        '200':
          description: Media found
          content:
            application/json:
              schema:
                $ref: './components/media-schemas.yaml#/components/schemas/MediaResponse'
    put:
      tags:
        - MediaCommand
      operationId: updateMedia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/media-schemas.yaml#/components/schemas/CreateMediaRequest'
      responses:
        202:
          description: UPDATED
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'
    delete:
      tags:
        - MediaCommand
      operationId: deleteMedia
      responses:
        200:
          description: DELETED
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'
    parameters:
      - name: mediaId
        in: path
        required: true
        schema:
          type: string
          format: uuid

  /api/media/{mediaId}/status:
    patch:
      tags:
        - MediaCommand
      operationId: changeMediaStatus
      summary: Change media status
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/media-schemas.yaml#/components/schemas/ChangeMediaStatusRequest'
      responses:
        '204':
          description: Status changed

## REVIEW
  /api/review:
    get:
      tags:
        - ReviewQuery
      operationId: findAllReview
      summary: Find all review
      responses:
        '200':
          description: List of review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './components/review-schemas.yaml#/components/schemas/ReviewResponse'

    post:
      tags:
        - ReviewCommand
      operationId: createReview
      summary: Create a review item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/review-schemas.yaml#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'

  /api/review/{reviewId}:
    get:
      tags:
        - ReviewQuery
      operationId: findReviewById
      summary: Find review by id
      responses:
        '200':
          description: Media found
          content:
            application/json:
              schema:
                $ref: './components/review-schemas.yaml#/components/schemas/ReviewResponse'
    put:
      tags:
        - ReviewCommand
      operationId: updateReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/review-schemas.yaml#/components/schemas/CreateReviewRequest'
      responses:
        202:
          description: UPDATED
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'
    delete:
      tags:
        - ReviewCommand
      operationId: deleteReview
      responses:
        200:
          description: DELETED
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'
    parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: string
          format: uuid
## USERs
  /api/user:
    post:
      tags:
        - User
      operationId: createUser
      summary: Create a user item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/shared-schemas.yaml#/components/schemas/CreateUserBody'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'

  /api/user/{userId}:
    get:
      tags:
        - User
      operationId: findUserById
      summary: Find user by id
      responses:
        '200':
          description: Media found
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/UserDTO'
    put:
      tags:
        - User
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/shared-schemas.yaml#/components/schemas/UserDTO'
      responses:
        202:
          description: UPDATED
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'
    delete:
      tags:
        - User
      operationId: deleteUser
      responses:
        200:
          description: DELETED
          content:
            application/json:
              schema:
                $ref: './components/shared-schemas.yaml#/components/schemas/IdentificationResponse'
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid